Resolved comments and separated non JIT specific code more clearly.
