+++ libs/execution_context/ebpf_core_jit.c
@@ -1,14 +1,14 @@
 // Copyright (c) eBPF for Windows contributors
 // SPDX-License-Identifier: MIT
 
-#include "ebpf_core_config.h"
+#include "ebpf_core_jit.h"
 #include "ebpf_core.h"
 #include "ebpf_program.h"
 #include "ebpf_maps.h"
 #include "ebpf_tracelog.h"
 #include "ebpf_serialize.h"
 
-// Forward declarations for external functions
+// Forward declarations for external functions.
 extern _Must_inspect_result_ ebpf_result_t ebpf_program_create_and_initialize(
     _In_ const ebpf_program_parameters_t* parameters, _Out_ ebpf_handle_t* program_handle);
