+++ libs/execution_context/kernel/execution_context_kernel.vcxproj
@@ -6,7 +6,7 @@
 <Project DefaultTargets="Build" ToolsVersion="12.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <ClCompile Include="..\ebpf_core.c" />
-    <ClCompile Include="..\ebpf_core_config.c" />
+    <ClCompile Include="..\ebpf_core_jit.c" />
     <ClCompile Include="..\ebpf_general_helpers.c" />
     <ClCompile Include="..\ebpf_link.c" />
     <ClCompile Include="..\ebpf_maps.c" />
@@ -16,7 +16,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\ebpf_core.h" />
-    <ClInclude Include="..\ebpf_core_config.h" />
+    <ClInclude Include="..\ebpf_core_jit.h" />
     <ClInclude Include="..\ebpf_link.h" />
     <ClInclude Include="..\ebpf_maps.h" />
     <ClInclude Include="..\ebpf_native.h" />
