+++ libs/execution_context/user/execution_context_user.vcxproj.filters
@@ -22,7 +22,7 @@
     <ClCompile Include="..\ebpf_core.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\ebpf_core_config.c">
+    <ClCompile Include="..\ebpf_core_jit.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\ebpf_general_helpers.c">
@@ -45,7 +45,7 @@
     <ClInclude Include="..\ebpf_core.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClCompile Include="..\ebpf_core_config.h">
+    <ClCompile Include="..\ebpf_core_jit.h">
       <Filter>Header Files</Filter>
     </ClCompile>
     <ClInclude Include="..\ebpf_link.h">
