+++ tests/end_to_end/end_to_end_jit.h
@@ -7,33 +7,18 @@ namespace ebpf {
 #include "net/if_ether.h"
 #include "net/ip.h"
 #include "net/udp.h"
-}; // namespace ebpf
+}; // namespace ebpf.
+
+#include "helpers.h"
 
 #define SAMPLE_PATH ""
 
-#define CONCAT(s1, s2) s1 s2
-#define DECLARE_TEST_CASE(_name, _group, _function, _suffix, _execution_type) \
-    TEST_CASE(CONCAT(_name, _suffix), _group) { _function(_execution_type); }
-#define DECLARE_NATIVE_TEST(_name, _group, _function) \
-    DECLARE_TEST_CASE(_name, _group, _function, "-native", EBPF_EXECUTION_NATIVE)
 #if !defined(CONFIG_BPF_JIT_DISABLED)
 #define DECLARE_JIT_TEST(_name, _group, _function) \
     DECLARE_TEST_CASE(_name, _group, _function, "-jit", EBPF_EXECUTION_JIT)
 #else
 #define DECLARE_JIT_TEST(_name, _group, _function)
 #endif
-#if !defined(CONFIG_BPF_INTERPRETER_DISABLED)
-#define DECLARE_INTERPRET_TEST(_name, _group, _function) \
-    DECLARE_TEST_CASE(_name, _group, _function, "-interpret", EBPF_EXECUTION_INTERPRET)
-#else
-#define DECLARE_INTERPRET_TEST(_name, _group, _function)
-#endif
-
-#define DECLARE_CGROUP_SOCK_ADDR_LOAD_TEST2(file, name, attach_type, name_suffix, file_suffix, execution_type) \
-    TEST_CASE("cgroup_sockaddr_load_test_" name "_" #attach_type "_" name_suffix, "[cgroup_sock_addr]")        \
-    {                                                                                                          \
-        cgroup_sock_addr_load_test(file file_suffix, name, attach_type, execution_type);                      \
-    }
 
 #if !defined(CONFIG_BPF_JIT_DISABLED)
 #define DECLARE_CGROUP_SOCK_ADDR_LOAD_JIT_TEST(file, name, attach_type) \
